fix bridge build issue
